+++ packages/web/design-system/src/ui/date-range-field/index.ts
@@ -1,0 +1,3 @@
+export { type DateRangeFieldProps as UIDateRangeFieldProps } from './dateRangeField.props.ts'
+export type { DateRangeFieldRange as UIDateRangeFieldRange } from './DateRangeField.vue'
+export { default as UIDateRangeField } from './DateRangeField.vue'